7903965: Fix Launching Tests from Context Menu
- merge the code of the dir and file config producers into a parent config producer
- prevent configuration creation when multiple files or directories are selected
- delete class config producer
- delete dir config producer

Use PsiClass/PsiMethod instead of PsiIdentifier:
- `findExactRunElement()`
- `nameForElement()`
- `getQuery`
